1.[VERSION]:0.8.4 2.[REMOVE & BUG FIX]:remove Y axis flip in projection matrix 3.[NOT USED]:stop using samplelight by area 4.[ADD]:start sample light by using random sphere direction with perturbation algo 5.[PERCISION ENHANCE]:in direct and indirect light, we use dvec3 instead of vec3

+++ include/scene/Scene.hpp
@@ -117,7 +117,7 @@ private:
   [[nodiscard]] std::tuple<Intersection, float> sampleLight();
 
   // sample the light by sphere angle and generate glm::vec3 direction
-  [[nodiscard]] std::tuple<glm::vec3, float>
+  [[nodiscard]] std::tuple<glm::dvec3, double>
   sampleLight(const glm::vec3 &shadingPoint);
 
   // Whitted Style Ray Tracing
